Simple support for texture, add IntVector2
+ IntVector2
+ Texture
- update mathutil structure

+++ Input/XboxController.cpp
@@ -70,7 +70,7 @@ void XboxController::updateThumbs(const XINPUT_STATE& xboxControllerState) {
 		x = rangeMapf(x, -32768.f, 32767.f, -1.f, 1.f);
 		y = rangeMapf(y, -32768.f, 32767.f, -1.f, 1.f);
 
-		float magnitude = sqrt(x*x + y*y), theta = atan2Degree(y, x);
+		float magnitude = sqrtf(x*x + y*y), theta = atan2Degree(y, x);
 		magnitude = rangeMapf(magnitude
 							  , THUMB_INNER_DEADZONE_FRACTION
 							  , THUMB_OUTER_DEADZONE_FRACTION
@@ -86,7 +86,7 @@ void XboxController::updateThumbs(const XINPUT_STATE& xboxControllerState) {
 		x = rangeMapf(x, -32768.f, 32767.f, -1.f, 1.f);
 		y = rangeMapf(y, -32768.f, 32767.f, -1.f, 1.f);
 
-		float magnitude = sqrt(x*x + y*y), theta = atan2Degree(y, x);
+		float magnitude = sqrtf(x*x + y*y), theta = atan2Degree(y, x);
 		magnitude = rangeMapf(magnitude
 							  , THUMB_INNER_DEADZONE_FRACTION
 							  , THUMB_OUTER_DEADZONE_FRACTION
